semhys-agents/requirements.txt

fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
google-cloud-discoveryengine==0.13.5
google-api-core==2.24.0
openai==1.59.5
google-generativeai==0.8.4
python-dotenv==1.0.1
requests==2.32.3
anthropic==0.42.0
